[frontend] Separate EC2 upload in a sub-controller
As we are adding the Azure upload we need to have each possible cloud
target separated in a different controller (in this case all will be
subclasses).
