+++ src/api/config/routes.rb
@@ -217,6 +217,11 @@ OBSApi::Application.routes.draw do
       end
       scope :ec2, as: :ec2 do
         resource :configuration, only: [:show, :update], controller: 'webui/cloud/ec2/configurations'
+        resource :upload, only: [:create], controller: 'webui/cloud/ec2/upload_jobs' do
+          new do
+            get ':project/:package/:repository/:arch/:filename', to: 'webui/cloud/ec2/upload_jobs#new', as: '', constraints: cons
+          end
+        end
       end
     end
